streamlit/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.7
botocore==1.35.7
cachetools==5.5.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
delta-kernel-rust-sharing-wrapper==0.1.0
delta-sharing==1.1.1
distlib==0.3.8
distro==1.9.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.65.0
grpcio==1.66.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.0
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.0.5
narwhals==1.5.5
nltk==3.9.1
nodeenv==1.9.1
numpy==2.1.0
openai==1.42.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pinecone==5.0.1
pinecone-cli==1.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
plotly==5.23.0
protobuf==4.25.4
protoc-gen-openapiv2==0.0.1
py==1.11.0
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
retry==0.9.2
rich==13.8.0
rpds-py==0.20.0
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.1
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
streamlit==1.38.0
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
watchdog==4.0.2
wheel==0.43.0
yarl==1.9.4
